app/requirements.txt
streamlit
Pillow
plotly